useMemo for assessment, metacache
